+++ tests/integration/test_dishln_abort_command.py
@@ -1,3 +1,5 @@
+import time
+
 import pytest
 import tango
 from ska_tango_base.commands import ResultCode
@@ -8,7 +10,9 @@ from tests.settings import (
     COMMAND_COMPLETED,
     DISH_LEAF_NODE_DEVICE,
     DISH_MASTER_DEVICE,
+    TIMEOUT,
     logger,
+    tear_down,
 )
 
 
@@ -20,7 +24,7 @@ def abort_on_dish_leaf_node(
     dev_factory = DevFactory()
     dish_leaf_node = dev_factory.get_device(DISH_LEAF_NODE_DEVICE)
     result_fp, _ = dish_leaf_node.AbortCommands()
-    assert result_fp[0] == ResultCode.OK
+    assert result_fp[0] == ResultCode.STARTED
 
 
 def abort_when_configured(
@@ -94,8 +98,73 @@ def abort_when_configured(
         f"Command ID: {unique_id_abort} Returned result: {result_abort}"
     )
 
-    assert result_abort == ResultCode.OK
+    assert result_abort == ResultCode.STARTED
 
+    group_callback["pointingState"].assert_change_event(
+        (PointingState.READY),
+        lookahead=6,
+    )
+    group_callback["dishMode"].assert_change_event(
+        (DishMode.STANDBY_FP),
+        lookahead=6,
+    )
+    dish_leaf_node.unsubscribe_event(DISHMODE_ID)
+    dish_leaf_node.unsubscribe_event(POINTINGSTATE_ID)
+    dish_leaf_node.unsubscribe_event(LRCR_ID)
+
+
+def abort_while_configuring(
+    tango_context,
+    dishln_name,
+    group_callback,
+    configure_input_str,
+):
+    logger.info(f"{tango_context}")
+    dev_factory = DevFactory()
+    dish_leaf_node = dev_factory.get_device(dishln_name)
+    dish_master = dev_factory.get_device(DISH_MASTER_DEVICE)
+    dish_master.SetDirectDishMode(DishMode.STANDBY_LP)
+    DISHMODE_ID = dish_leaf_node.subscribe_event(
+        "dishMode",
+        tango.EventType.CHANGE_EVENT,
+        group_callback["dishMode"],
+    )
+    POINTINGSTATE_ID = dish_leaf_node.subscribe_event(
+        "pointingState",
+        tango.EventType.CHANGE_EVENT,
+        group_callback["pointingState"],
+    )
+
+    group_callback["dishMode"].assert_change_event(
+        (DishMode.STANDBY_LP),
+        lookahead=2,
+    )
+
+    result_fp, unique_id_fp = dish_leaf_node.SetStandbyFPMode()
+    group_callback["dishMode"].assert_change_event(
+        (DishMode.STANDBY_FP),
+        lookahead=6,
+    )
+
+    assert result_fp[0] == ResultCode.QUEUED
+    LRCR_ID = dish_leaf_node.subscribe_event(
+        "longRunningCommandResult",
+        tango.EventType.CHANGE_EVENT,
+        group_callback["longRunningCommandResult"],
+    )
+
+    group_callback["longRunningCommandResult"].assert_change_event(
+        (unique_id_fp[0], COMMAND_COMPLETED),
+        lookahead=2,
+    )
+
+    result_config, unique_id_config = dish_leaf_node.Configure(
+        configure_input_str
+    )
+    assert result_config[0] == ResultCode.QUEUED
+    logger.info(
+        f"Command ID: {unique_id_config} Returned result: {result_config}"
+    )
     group_callback["pointingState"].assert_change_event(
         (PointingState.READY),
         lookahead=6,
@@ -104,9 +173,30 @@ def abort_when_configured(
         (DishMode.OPERATE),
         lookahead=6,
     )
+
+    result_abort, unique_id_abort = dish_leaf_node.AbortCommands()
+    logger.info(
+        f"Command ID: {unique_id_abort} Returned result: {result_abort}"
+    )
+
+    assert result_abort == ResultCode.STARTED
+
+    group_callback["dishMode"].assert_change_event(
+        (DishMode.STANDBY_FP),
+        lookahead=6,
+    )
+    assert (
+        dish_leaf_node.read_attribute("pointingState").value
+        == PointingState.READY
+    )
+
+    assert is_configure_aborted(dish_leaf_node, unique_id_config[0])
+
     dish_leaf_node.unsubscribe_event(DISHMODE_ID)
     dish_leaf_node.unsubscribe_event(POINTINGSTATE_ID)
     dish_leaf_node.unsubscribe_event(LRCR_ID)
+
+    tear_down(dish_leaf_node, dish_master, group_callback)
 
 
 @pytest.mark.post_deployment
@@ -120,10 +210,39 @@ def test_abort_command(tango_context, group_callback):
 
 @pytest.mark.post_deployment
 @pytest.mark.SKA_mid
-def test_abort_when_configure(tango_context, group_callback, json_factory):
+def test_abort_after_configure(tango_context, group_callback, json_factory):
     abort_when_configured(
         tango_context,
         DISH_LEAF_NODE_DEVICE,
         group_callback,
         json_factory("dishleafnode_configure"),
     )
+
+
+@pytest.mark.post_deployment
+@pytest.mark.SKA_mid
+def test_abort_while_configuring(tango_context, group_callback, json_factory):
+    abort_while_configuring(
+        tango_context,
+        DISH_LEAF_NODE_DEVICE,
+        group_callback,
+        json_factory("dishleafnode_configure"),
+    )
+
+
+def is_configure_aborted(dish_leaf_node, unique_id_config) -> bool:
+    """
+    This method checks if Configure command is Aborted for given command id.
+    """
+    start_time = time.time()
+    elapsed_time = 0
+    while elapsed_time < TIMEOUT:
+        lrcs_value = dish_leaf_node.longRunningCommandStatus
+        lrcs_iterator = iter(lrcs_value)
+        for value in lrcs_iterator:
+            if value == unique_id_config:
+                if next(lrcs_iterator) == "ABORTED":
+                    return True
+        time.sleep(0.1)
+        elapsed_time = time.time() - start_time
+    return False