+++ tests/unit/test_device_unresponsive.py
@@ -2,7 +2,7 @@ import pytest
 from ska_tmc_common.exceptions import DeviceUnresponsive
 
 from ska_tmc_dishleafnode.dish_leaf_node import DishLeafNode
-from tests.settings import DISH_MASTER_DEVICE, logger, wait_for_unresponsive
+from tests.settings import DISH_MASTER_DEVICE, logger
 
 
 @pytest.fixture()
@@ -22,7 +22,7 @@ def test_abort_command_fail_check_allowed_with_device_unresponsive(
     tango_context, cm
 ):
     logger.info("%s", tango_context)
-    wait_for_unresponsive(cm)
+    cm.get_device().update_unresponsive(True, "Not available")
     with pytest.raises(
         DeviceUnresponsive, match=f"{DISH_MASTER_DEVICE} not available"
     ):
@@ -31,6 +31,6 @@ def test_abort_command_fail_check_allowed_with_device_unresponsive(
 
 def test_trackloadstaticoff_command_not_allowed(tango_context, cm):
     """Test the command not allowed when the device is unresponsive."""
-    wait_for_unresponsive(cm)
+    cm.get_device().update_unresponsive(True, "Not available")
     with pytest.raises(DeviceUnresponsive):
         cm.is_trackloadstaticoff_allowed()